poll: Fix size types in ssh_event_free()
src/poll.c:1024:9: error: assuming signed overflow does not occur when
    simplifying conditional to constant [-Werror=strict-overflow]

(cherry picked from commit 11d87238b86732d53378e8acf792838914a38f14)

+++ src/poll.c
@@ -1015,7 +1015,7 @@ int ssh_event_remove_connector(ssh_event event, ssh_connector connector){
  */
 void ssh_event_free(ssh_event event)
 {
-    int used, i;
+    size_t used, i;
     ssh_poll_handle p;
 
     if(event == NULL) {
